fix: displaced feedback button
