+++ packages/shared/src/components/squads/SquadHeaderBar.tsx
@@ -8,7 +8,12 @@ import useContextMenu from '../../hooks/useContextMenu';
 import SquadMemberShortList, {
   SquadMemberShortListProps,
 } from './SquadMemberShortList';
-import { IconSize } from '../Icon';
+import { LazyModal } from '../modals/common/types';
+import { useLazyModal } from '../../hooks/useLazyModal';
+import { squadFeedback } from '../../lib/constants';
+import FeedbackIcon from '../icons/Feedback';
+import { useAuthContext } from '../../contexts/AuthContext';
+import useSidebarRendered from '../../hooks/useSidebarRendered';
 
 interface SquadHeaderBarProps
   extends SquadMemberShortListProps,
@@ -21,16 +26,25 @@ export function SquadHeaderBar({
   members,
   memberCount,
   className,
-  onNewSquadPost,
   ...props
 }: SquadHeaderBarProps): ReactElement {
+  const { user } = useAuthContext();
+  const { sidebarRendered } = useSidebarRendered();
   const { onMenuClick } = useContextMenu({ id: 'squad-menu-context' });
+  const { openModal } = useLazyModal();
+  const openSquadInviteModal = () =>
+    openModal({
+      type: LazyModal.SquadInvite,
+      props: {
+        initialSquad: squad,
+      },
+    });
 
   return (
     <div
       {...props}
       className={classNames(
-        'flex flex-row flex-wrap justify-center items-center gap-4 w-full',
+        'flex laptop:flex-row flex-col justify-center items-center gap-4 w-full',
         className,
       )}
     >
@@ -39,20 +53,30 @@ export function SquadHeaderBar({
         members={members}
         memberCount={memberCount}
       />
-      <SimpleTooltip placement="top" content="Squad options">
+      <span className="flex flex-row gap-4">
+        <Button className="btn-secondary" onClick={openSquadInviteModal}>
+          Invite
+        </Button>
         <Button
-          className="tablet:order-2 btn btn-secondary"
-          icon={<MenuIcon size={IconSize.Small} />}
-          onClick={onMenuClick}
-          aria-label="Squad options"
-        />
-      </SimpleTooltip>
-      <Button
-        className="w-full mobileL:max-w-[18rem] tablet:w-fit btn btn-primary"
-        onClick={onNewSquadPost}
-      >
-        Create new post
-      </Button>
+          tag="a"
+          target="_blank"
+          rel="noopener noreferrer"
+          href={`${squadFeedback}#user_id=${user.id}&squad_id=${squad.id}`}
+          className="btn-secondary"
+          icon={<FeedbackIcon aria-label="squad-feedback-icon" />}
+          aria-label="squad-feedback"
+        >
+          {sidebarRendered ? 'Feedback' : ''}
+        </Button>
+        <SimpleTooltip placement="top" content="Squad options">
+          <Button
+            className="btn-secondary"
+            icon={<MenuIcon />}
+            onClick={onMenuClick}
+            aria-label="Squad options"
+          />
+        </SimpleTooltip>
+      </span>
       <SquadHeaderMenu squad={squad} />
     </div>
   );
